Started on tryout the cv matcher dashboard page
The idea is to have a page where you have a code editor where you
can edit the send json data directly.
The editor will will have a json schema for the CV so you know what's
possible and see what's wrong quickly.

This page should also return the match result so you know how and to
wich profile the cv was matched.