introduce routeBuilder.R to define more route information
Currently this doesn't have much added value.

i'm planning to use this for the openAPI route to provide the output with more information about the routes defined what should improve the developer experiance.

as a bonus every defined with the routeBuilder.R now also has a explainer comment above itself so it's also easier for someone else to understand what every route is doing