Simplify authentication
I was looking at the server setup and the authentication logic on a lot
of server side things are way easier than this app.
So if all things on the server don't require complex rolling keys etc is
it realy needed this app needs it?
As i'm trying to keep things simple i have chosen to simplify the
authentication method by a lot.