Crypto: add Encrypt(Reader&Writer)
Add a encryption reader and writer that encrypt data in chunks.
By adding this we can encrypt large amounds of data without having it
all in memory.

This is mainly in preperation for the backup support so we can safely
encrypt large amounds to a file without having to worry about OOM
errors.