The start of a new PDF generator
Written in dart with flutter like widgets this pdf generator will
generate the CV PDFs very quick and with good looking results.

This is just the start and currently it just renders a static data.
Somewhere in the futhure we will swapout the static data and make
everything dynamic so we can use this in the base project for the CV PDF
generation.

I've done some quick benchmarks and on my machine (i7 9th gen) it takes
about 25ms to generate a PDF file when the dart code is compiled.